feature/crawl_data: crawl team and team rank

+++ Services/api/src/migrations/20230702105424_add_driver_rank.ts
@@ -16,7 +16,7 @@ export async function up(knex: Knex): Promise<void> {
       .onUpdate("CASCADE")
       .onDelete("NO ACTION");
     table.integer("position");
-    table.integer("points");
+    table.decimal("points");
     table.integer("completed_laps");
     table.string("finish_time");
     table.integer("start_position");
